+++ instagram_to_discord/sites/instagram/converter_instagram_url.py
@@ -1,9 +1,6 @@
 # __a にアクセスする関数
 import re
 
-from debug import DEBUG
-
-
 def convert_instagram_url_to_a(url):
     # https://www.instagram.com/p/CJ8u5PCH-WG/?utm_source=ig_web_copy_link
     m = re.match(r"https://www.instagram.com/(p|reel)/([^/])+", url)
@@ -31,8 +28,6 @@ def extract_url(line: str):
 
 def instagram_extract_from_content(content: str):
     for line in content.split("\n"):
-        if DEBUG:
-            print("[debug extract from..] line: {}".format(line))
         base_url = extract_url(line)
         if base_url:
             return base_url
